PR review modifactions
